wallet: split itest RenameAccount tests
